move wait for completion check and fire execution inside JobLauncher
