+++ server/src/main/java/io/crate/planner/statement/CopyFromPlan.java
@@ -131,8 +131,8 @@ public final class CopyFromPlan implements Plan {
         JobLauncher jobLauncher = dependencies.phasesTaskFactory()
             .create(plannerContext.jobId(), List.of(nodeOpTree));
 
-        CopyPlan.execute(consumer, plannerContext.transactionContext(), jobLauncher::execute,
-                         boundedCopyFrom.settings().getAsBoolean("wait_for_completion", true));
+        JobLauncher.execute(consumer, plannerContext.transactionContext(), jobLauncher::execute,
+                     boundedCopyFrom.settings().getAsBoolean("wait_for_completion", true));
     }
 
     @VisibleForTesting
